chore(app): prune legacy JSON-RPC parser; introduce shared types; update imports and tests; format
