+++ packages/app/src/preload.ts
@@ -1,5 +1,6 @@
 import { contextBridge, ipcRenderer } from 'electron';
 import { version } from './version';
+import type { StateAtArguments, StateAtResult } from './types';
 
 /**
  * Preload script executed in an isolated context.
@@ -8,18 +9,6 @@ import { version } from './version';
  */
 
 // Minimal, typed-safe preload bridge. Extend via adapters/host APIs only.
-interface StateAtArguments {
-  asOf: string;
-  scenario?: string;
-  confidence?: number;
-}
-interface StateAtResult {
-  asOf: string;
-  scenario: string | null;
-  confidence: number | null;
-  nodes: number;
-  edges: number;
-}
 
 contextBridge.exposeInMainWorld('aideon', {
   version,
